simple_filters_in_action/requirements.txt
Flask==1.1
requests==2.24
Pillow==7.1
